move SearchTypes to user-group-response file
