+++ packages/app/src/interfaces/user-group-response.ts
@@ -37,3 +37,11 @@ export type SelectableUserChildGroupsResult = {
 export type AncestorUserGroupsResult = {
   ancestorUserGroups: IUserGroupHasId[],
 }
+
+export const SearchTypes = {
+  FORWARD: 'forward',
+  PARTIAL: 'partial',
+  BACKWORD: 'backword',
+} as const;
+
+export type SearchType = typeof SearchTypes[keyof typeof SearchTypes];
