--- packages/app/src/interfaces/user-group.ts
@@ -1,7 +1,0 @@
-export const SearchTypes = {
-  FORWARD: 'forward',
-  PARTIAL: 'partial',
-  BACKWORD: 'backword',
-} as const;
-
-export type SearchType = typeof SearchTypes[keyof typeof SearchTypes];
